pyproject.toml
[project]
name = "stroke-prediction"
version = "1.0.0"
description = "Machine Learning-Based Stroke Risk Prediction"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Data Science Team" }
]
keywords = ["machine-learning", "stroke-prediction", "healthcare", "data-science"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "imbalanced-learn>=0.11.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "shap>=0.45.0",
    "joblib>=1.3.0",
    "python-docx>=1.2.0",
    "markdown>=3.9",
    "ipykernel>=6.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["utils"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
